Prevented taggatron from clobbering an existing &tags value
Updated taggatron#SetTags() to take into account the existing value of
&tags, ie: if the tag file is already picket up by the current &tags
then tags are not updated. Otherwise, the path to the tag file is
appended to the end of the current &tags value.

Values added to tagdefaults are appended to &tags upon plugin
initialisation. Thus, allowing users to define not just individual
tagfiles but also custom search patterns.

Updated taggatron#CreateTags() to make use of the taggatron#SetTags(),
instead of manipulating &tags value directly.